myproject/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
pytest>=7.4.0
pytest-django>=4.5.2
allure-pytest>=2.13.2
requests>=2.31.0